test: piece and coords
